Update concurrency to 1 per machine
This means only one sample_init will happen per machine, which is an IO
heavy operation.

Actual running of the evals should retain the concurrency and run once
per proxmox instance.

This also fixes a clash where "proxmox" was the name of the sandbox
concurrency also. So now we will have "proxmox" limiting to the number
of instances (the normal --max_sandboxes), and "proxmox-(machine name)",
with the latter appearing once for each machine.